Marketplace Thank You: Update the page moving each product type logic into their own hook
